FIX Do not warn when using polars DataFrames in DecisionBoundaryDisplay.from_estimator (#28718)
