+++ sklearn/preprocessing/_function_transformer.py
@@ -4,7 +4,10 @@ import numpy as np
 
 from ..base import BaseEstimator, TransformerMixin, _fit_context
 from ..utils._param_validation import StrOptions
-from ..utils._set_output import ADAPTERS_MANAGER, _get_output_config
+from ..utils._set_output import (
+    _get_adapter_from_container,
+    _get_output_config,
+)
 from ..utils.metaestimators import available_if
 from ..utils.validation import (
     _allclose_dense_sparse,
@@ -14,24 +17,6 @@ from ..utils.validation import (
     _is_polars_df,
     check_array,
 )
-
-
-def _get_adapter_from_container(container):
-    """Get the adapter that nows how to handle such container.
-
-    See :class:`sklearn.utils._set_output.ContainerAdapterProtocol` for more
-    details.
-    """
-    module_name = container.__class__.__module__.split(".")[0]
-    try:
-        return ADAPTERS_MANAGER.adapters[module_name]
-    except KeyError as exc:
-        available_adapters = list(ADAPTERS_MANAGER.adapters.keys())
-        raise ValueError(
-            "The container does not have a registered adapter in scikit-learn. "
-            f"Available adapters are: {available_adapters} while the container "
-            f"provided is: {container!r}."
-        ) from exc
 
 
 def _identity(X):
