+++ sklearn/utils/_set_output.py
@@ -197,6 +197,24 @@ ADAPTERS_MANAGER.register(PandasAdapter())
 ADAPTERS_MANAGER.register(PolarsAdapter())
 
 
+def _get_adapter_from_container(container):
+    """Get the adapter that knows how to handle such container.
+
+    See :class:`sklearn.utils._set_output.ContainerAdapterProtocol` for more
+    details.
+    """
+    module_name = container.__class__.__module__.split(".")[0]
+    try:
+        return ADAPTERS_MANAGER.adapters[module_name]
+    except KeyError as exc:
+        available_adapters = list(ADAPTERS_MANAGER.adapters.keys())
+        raise ValueError(
+            "The container does not have a registered adapter in scikit-learn. "
+            f"Available adapters are: {available_adapters} while the container "
+            f"provided is: {container!r}."
+        ) from exc
+
+
 def _get_container_adapter(method, estimator=None):
     """Get container adapter."""
     dense_config = _get_output_config(method, estimator)["dense"]
